Add GridTableApi.getSelectedRows.
